Fixed some Package Me table definitions.

+++ db/package_me/tables/audit_trail/at.object_snapshot_links.rb
@@ -7,7 +7,7 @@ table :object_snapshot_links, sql_schema: :at do
   col :code, len: 32, required: true
   col :parent_snapshot_id, :ref => 'at.object_snapshots', required: true
   col :child_snapshot_id, :ref => 'at.object_snapshots', required: true
-  col :ordinal, type: :int
+  col :ordinal, type: :byte
   col :status, type: :byte, required: true
   col :modified_at, required: true, default: { function: :getdate }, index: { clustered: true }
   col :modified_by, required: true
